Add lazy and eager loading options for embeddable scripts
This update modifies the .eleventy.js file to allow for 'lazy' or 'eager' loading of embeddable scripts for forms and meetings. It provides a more efficient way of loading these external resources by checking if a loadingMode is set in the plugin options and throws an error if the loadingMode isn't one of 'default', 'lazy', or 'eager'.